Django Framework/Milestone 1/Blog Post/requirements.txt
asgiref==3.11.0
Django==6.0
django-debug-toolbar==6.1.0
psycopg2==2.9.11
sqlparse==0.5.5
tzdata==2025.3
